requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
greenlet==3.2.1
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pillow==11.2.1
scikit-image==0.25.2
scipy==1.15.2
SQLAlchemy==2.0.40
tifffile==2025.3.30
typing_extensions==4.13.2
Werkzeug==3.1.3
